+++ sema/type_checker.h
@@ -50,6 +50,8 @@ public:
     void visit(AsmStmt* node) override;
     void visit(ThreadCreateExpr* node) override;
     void visit(ThreadJoinStmt* node) override;
+    void visit(ThrowStmt* node) override;
+    void visit(TryStmt* node) override;
 
 private:
     // Symbol table: maps name -> type